+++ opensrp-core/src/main/java/org/opensrp/service/ReferralsReportService.java
@@ -59,7 +59,7 @@ public class ReferralsReportService {
         for (ClientRegistrationReason clientRegistrationReason : clientRegistrationReasons) {
             RegistrationReasonsReportDTO registrationReasonsReportDTO = new RegistrationReasonsReportDTO();
             registrationReasonsReportDTO.setRegistrationReason(clientRegistrationReason.getDescEn());
-            int totalMale = 0, totalFemale = 0;
+            long totalMale = 0, totalFemale = 0;
 
             String lessThan1year = generateRegistationReportSql(firstDateOfTheMonth.withDayOfMonth(1).toString(), currentDate, getDateByYearString(-1), currentDate, clientRegistrationReason.getRegistrationId());
